Rename AsyncRegistryServer/ to SystemMonitor/AsyncRegistry
