+++ arangod/CMakeLists.txt
@@ -103,7 +103,6 @@ get_target_property(IRESEARCH_INCLUDE
 include(arangoserver.cmake)
 add_subdirectory(Agency)
 add_subdirectory(Aql)
-add_subdirectory(AsyncRegistryServer)
 add_subdirectory(Cache)
 add_subdirectory(Cluster)
 add_subdirectory(ClusterEngine)
@@ -118,6 +117,7 @@ add_subdirectory(RestHandler)
 add_subdirectory(RestServer)
 add_subdirectory(RocksDBEngine)
 add_subdirectory(StorageEngine)
+add_subdirectory(SystemMonitor)
 add_subdirectory(Utils)
 if (USE_V8)
   add_subdirectory(V8Server)
