+++ arangod/RestServer/arangod_includes.h
@@ -51,7 +51,7 @@
 #include "Aql/AqlFunctionFeature.h"
 #include "Aql/OptimizerRulesFeature.h"
 #include "Aql/QueryInfoLoggerFeature.h"
-#include "AsyncRegistryServer/Feature.h"
+#include "SystemMonitor/AsyncRegistry/Feature.h"
 #include "Basics/ArangoGlobalContext.h"
 #include "Basics/FileUtils.h"
 #include "Basics/directories.h"
