+++ arangod/SystemMonitor/AsyncRegistry/Stacktrace/depth_first.h
