+++ arangod/SystemMonitor/AsyncRegistry/Stacktrace/forest.h
